utils/requirements.txt
peft==0.12.0
botocore
Pillow==11.2.1
accelerate>=0.26.0
qwen_vl_utils
torch==2.4.0
torchvision==0.19.0
transformers==4.46.0